+++ api/src/main/java/io/sillysillyman/api/auth/config/JwtConfig.java
@@ -1,4 +1,4 @@
-package io.sillysillyman.api.controller.auth.config;
+package io.sillysillyman.api.auth.config;
 
 import io.jsonwebtoken.security.Keys;
 import io.sillysillyman.core.auth.properties.JwtProperties;
