+++ api/src/main/java/io/sillysillyman/api/auth/handler/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package io.sillysillyman.api.controller.auth.handler;
+package io.sillysillyman.api.auth.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.sillysillyman.api.common.dto.ErrorResponse;
@@ -10,32 +10,32 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
-@Slf4j(topic = "JwtAccessDeniedHandler")
+@Slf4j(topic = "JwtAuthenticationEntryPoint")
 @Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
     private final ObjectMapper objectMapper;
 
     @Override
-    public void handle(
+    public void commence(
         HttpServletRequest request,
         HttpServletResponse response,
-        AccessDeniedException accessDeniedException
+        AuthenticationException authException
     ) throws IOException, ServletException {
-        log.error("Access denied error: {}", accessDeniedException.getMessage());
+        log.error("Unauthorized error: {}", authException.getMessage());
 
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
 
         ErrorResponse errorResponse = ErrorResponse.of(
-            accessDeniedException.getMessage(),
-            HttpStatus.FORBIDDEN.value(),
-            HttpStatus.FORBIDDEN.name()
+            authException.getMessage(),
+            HttpStatus.UNAUTHORIZED.value(),
+            HttpStatus.UNAUTHORIZED.name()
         );
 
         objectMapper.writeValue(response.getOutputStream(), errorResponse);
